replace camera width/height with single resolution + add env tutial pynb

+++ animalai/animalai/envs/environment.py
@@ -39,15 +39,12 @@ class AnimalAIEnvironment(UnityEnvironment):
         play: bool = False,
         arenas_configurations: ArenaConfig = None,
         inference: bool = False,
-        camera_width: int = None,
-        camera_height: int = None,
+        resolution: int = None,
         grayscale: bool = False,
         side_channels: Optional[List[SideChannel]] = None,
     ):
 
-        args = self.executable_args(
-            n_arenas, play, camera_height, camera_width, grayscale
-        )
+        args = self.executable_args(n_arenas, play, resolution, grayscale)
         self.play = play
         self.inference = inference
         self.timeout = 10 if play else 60
@@ -126,7 +123,8 @@ class AnimalAIEnvironment(UnityEnvironment):
     def close(self):
         if self.play:
             self.communicator.close()
-            self.proc1.kill()
+            if self.proc1:
+                self.proc1.kill()
         else:
             super().close()
 
@@ -134,8 +132,7 @@ class AnimalAIEnvironment(UnityEnvironment):
     def executable_args(
         n_arenas: int = 1,
         play: bool = False,
-        camera_height: int = 84,
-        camera_width: int = 84,
+        resolution: int = 84,
         grayscale: bool = False,
     ) -> List[str]:
         args = ["--playerMode"]
@@ -145,12 +142,9 @@ class AnimalAIEnvironment(UnityEnvironment):
             args.append("0")
         args.append("--numberOfArenas")
         args.append(str(n_arenas))
-        if camera_width:
-            args.append("--cameraWidth")
-            args.append(str(camera_width))
-        if camera_height:
-            args.append("--cameraHeight")
-            args.append(str(camera_height))
+        if resolution:
+            args.append("--resolution")
+            args.append(str(resolution))
         if grayscale:
             args.append("--grayscale")
         return args
